Added new exception, some improvements

+++ src/backend/src/main/java/iRide/repository/InstructorCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InstructorCategoryRepository extends JpaRepository<InstructorCategory, Long> {
+public interface InstructorCategoryRepository extends JpaRepository<InstructorCategory, Integer> {
 }
