+++ src/backend/src/main/java/iRide/repository/InstructorVehicleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InstructorVehicleRepository extends JpaRepository<InstructorVehicle, Long> {
+public interface InstructorVehicleRepository extends JpaRepository<InstructorVehicle, Integer> {
 }
